Organize x86 specific code into a separate create
