+++ Cargo.toml
@@ -8,16 +8,18 @@ authors = [ "robert.w.gries@gmail.com" ]
 crate-type = ["staticlib"]
 
 [dependencies]
-bitflags = "0.7.0"
 compiler_builtins = { git = "https://github.com/rust-lang-nursery/compiler-builtins" }
 multiboot2 = "0.3.1"
-rlibc = "1.0.0"
-spin = "0.3.4"
-volatile = "0.1.0"
 
 [dependencies.x86]
 version = "0.8.0"
 default-features = false
+
+[target.'cfg(target_arch = "x86")'.dependencies]
+arch_x86 = { path = "arch/x86" }
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
+arch_x86 = { path = "arch/x86" }
 
 # The development profile, used for `cargo build`.
 [profile.dev]
