+++ arch/x86/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "arch_x86"
+version = "0.1.0"
+authors = [ "robert.w.gries@gmail.com" ]
+
+[dependencies]
+bitflags = "0.7.0"
+multiboot2 = "0.3.1"
+once = "0.3.2"
+rlibc = "1.0.0"
+spin = "0.3.4"
+volatile = "0.1.0"
+
+[dependencies.x86]
+version = "0.8.0"
+default-features = false
