+++ arch/x86/asm/check-common.S
