+++ arch/x86/asm/startup-common.S
