+++ arch/x86/src/console.rs
@@ -1,0 +1,18 @@
+use core::fmt::{self, Write};
+use spin::Mutex;
+
+use device::vga::VGA;
+
+pub static CONSOLE: Mutex<Console> = Mutex::new(Console);
+
+pub struct Console;
+
+impl Write for Console {
+    fn write_str(&mut self, s: &str) -> Result<(), fmt::Error> {
+        VGA.lock().write_str(s)
+    }
+}
+
+pub fn init() {
+    VGA.lock().clear_screen();
+}
