+++ arch/x86/src/device/mod.rs
@@ -1,0 +1,1 @@
+pub mod vga;
