+++ arch/x86/src/memory/paging/mod.rs
@@ -135,9 +135,11 @@ pub fn remap_the_kernel<A>(allocator: &mut A, boot_info: &BootInformation)
             assert!(section.addr as usize % PAGE_SIZE == 0,
                     "sections need to be page aligned");
 
+/*
             println!("mapping section at addr: {:#x}, size: {:#x}",
                      section.addr,
                      section.size);
+*/
 
             let flags = EntryFlags::from_elf_section_flags(section);
 
@@ -163,5 +165,5 @@ pub fn remap_the_kernel<A>(allocator: &mut A, boot_info: &BootInformation)
     // turn the old p4 page into a guard page
     let old_p4_page = Page::containing_address(old_table.p4_frame.start_address());
     active_table.unmap(old_p4_page, allocator);
-    println!("guard page at {:#x}", old_p4_page.start_address());
+    //println!("guard page at {:#x}", old_p4_page.start_address());
 }
